More install steps automated
Automatically create Missions\Kaukasus Insurgency folder if not exists, including all it's child directories.

Copy JSON.lua over or add if not exists.

First version of opening up DCS miz file, and updating the paths in the dictionary file to point to current path where update is being installed for demo missions. Not finished as we still need to update the zip stream and then save it.